Revamp since the SAMLInitializer and unmarshall method was moved to saml common repo
