+++ components/org.wso2.carbon.identity.oauth/src/test/java/org/wso2/carbon/identity/oauth2/token/handlers/grant/saml/SAML2BearerGrantHandlerTest.java
@@ -24,6 +24,7 @@ import org.joda.time.DateTime;
 import org.junit.runner.RunWith;
 import org.mockito.Matchers;
 import org.mockito.Mock;
+import org.opensaml.core.xml.io.UnmarshallingException;
 import org.opensaml.saml.saml2.core.Assertion;
 import org.opensaml.saml.saml2.core.NameID;
 import org.opensaml.saml.saml2.core.Subject;
@@ -54,7 +55,6 @@ import org.wso2.carbon.identity.application.common.model.ServiceProvider;
 import org.wso2.carbon.identity.application.common.util.IdentityApplicationConstants;
 import org.wso2.carbon.identity.application.common.util.IdentityApplicationManagementUtil;
 import org.wso2.carbon.identity.application.mgt.ApplicationManagementService;
-import org.wso2.carbon.identity.base.IdentityException;
 import org.wso2.carbon.identity.core.model.SAMLSSOServiceProviderDO;
 import org.wso2.carbon.identity.core.persistence.IdentityPersistenceManager;
 import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
@@ -71,6 +71,8 @@ import org.wso2.carbon.identity.oauth2.token.OAuthTokenReqMessageContext;
 import org.wso2.carbon.identity.oauth2.token.OauthTokenIssuer;
 import org.wso2.carbon.identity.oauth2.token.OauthTokenIssuerImpl;
 import org.wso2.carbon.identity.oauth2.util.OAuth2Util;
+import org.wso2.carbon.identity.saml.common.util.UnmarshallUtils;
+import org.wso2.carbon.identity.saml.common.util.exception.IdentityUnmarshallingException;
 import org.wso2.carbon.identity.sso.saml.SSOServiceProviderConfigManager;
 import org.wso2.carbon.identity.sso.saml.dto.SAMLSSOAuthnReqDTO;
 import org.wso2.carbon.identity.sso.saml.util.SAMLSSOUtil;
@@ -112,7 +114,7 @@ import static org.testng.Assert.fail;
 @PrepareForTest({IdentityUtil.class, IdentityTenantUtil.class, IdentityProviderManager.class, MultitenantUtils.class,
         IdentityApplicationManagementUtil.class, OAuthServerConfiguration.class, SSOServiceProviderConfigManager.class,
         SAML2BearerGrantHandler.class, OAuthComponentServiceHolder.class, OAuth2ServiceComponentHolder.class,
-        OAuth2Util.class,IdentityPersistenceManager.class, SignatureValidator.class})
+        OAuth2Util.class,IdentityPersistenceManager.class, SignatureValidator.class, UnmarshallUtils.class})
 public class SAML2BearerGrantHandlerTest extends PowerMockIdentityBaseTest {
 
     public static final String[] SCOPE_ARRAY = {"scope1"};
@@ -164,6 +166,7 @@ public class SAML2BearerGrantHandlerTest extends PowerMockIdentityBaseTest {
         );
         mockStatic(OAuthServerConfiguration.class);
         mockStatic(IdentityUtil.class);
+        mockStatic(UnmarshallUtils.class);
         when(OAuthServerConfiguration.getInstance()).thenReturn(oAuthServerConfiguration);
         when(oAuthServerConfiguration.getIdentityOauthTokenIssuer()).thenReturn(oauthIssuer);
         when(oAuthServerConfiguration.getPersistenceProcessor()).thenReturn(persistenceProcessor);
@@ -224,7 +227,7 @@ public class SAML2BearerGrantHandlerTest extends PowerMockIdentityBaseTest {
         DateTime validOnOrAfter = new DateTime(System.currentTimeMillis() + 10000000L);
         DateTime expiredOnOrAfter = new DateTime(System.currentTimeMillis() - 10000000L);
         return new Object[][] {
-                { validOnOrAfter, "LOCAL", true, true, TestConstants.OAUTH2_TOKEN_EP, TestConstants.LOACALHOST_DOMAIN, new IdentityException("Error"), "Error while unmashalling"},
+                { validOnOrAfter, "LOCAL", true, true, TestConstants.OAUTH2_TOKEN_EP, TestConstants.LOACALHOST_DOMAIN, new IdentityUnmarshallingException("Error"), "Error while unmashalling"},
                 { validOnOrAfter, "FED", true, true, TestConstants.OAUTH2_TOKEN_EP, TestConstants.LOACALHOST_DOMAIN, new IdentityProviderManagementException("Error"), "Error while retrieving identity provider"},
                 { validOnOrAfter, "FED", true, true, TestConstants.OAUTH2_TOKEN_EP, TestConstants.LOACALHOST_DOMAIN, new SignatureException(), "Error while validating the signature"},
                 { validOnOrAfter, "LOCAL", true, true, TestConstants.OAUTH2_TOKEN_EP, TestConstants.LOACALHOST_DOMAIN, new IdentityApplicationManagementException("Error"), "Error while retrieving service provider"},
@@ -272,8 +275,8 @@ public class SAML2BearerGrantHandlerTest extends PowerMockIdentityBaseTest {
         if (e instanceof IdentityProviderManagementException) {
             when(identityProviderManager.getIdPByAuthenticatorPropertyValue(anyString(), anyString(), anyString(),
                     anyString(), anyBoolean())).thenThrow(e);
-        } else if (e instanceof IdentityException) {
-            when(IdentityUtil.unmarshall(anyString())).thenThrow(e);
+        } else if (e instanceof IdentityUnmarshallingException) {
+            when(UnmarshallUtils.unmarshall(anyString())).thenThrow(e);
         } else if (e instanceof SignatureException) {
             PowerMockito.mockStatic(SignatureValidator.class);
             PowerMockito.doThrow(e)
@@ -464,7 +467,7 @@ public class SAML2BearerGrantHandlerTest extends PowerMockIdentityBaseTest {
         String assertionString = SAMLSSOUtil.marshall(assertion);
         assertionString = new String(Base64.encodeBase64(assertionString.getBytes(Charsets.UTF_8)), Charsets.UTF_8);
         oAuth2AccessTokenReqDTO.setAssertion(assertionString);
-        when(IdentityUtil.unmarshall(anyString())).thenReturn(assertion);
+        when(UnmarshallUtils.unmarshall(anyString())).thenReturn(assertion);
         when(IdentityUtil.isTokenLoggable(anyString())).thenReturn(true);
     }
